+++ src/stream/memory.rs
@@ -31,8 +31,8 @@ impl SeekStream for MemoryStream {
         Ok(self.cursor.stream_position()?)
     }
 
-    fn len(&self) -> BinaryResult<usize> {
-        Ok(self.cursor.get_ref().len())
+    fn len(&self) -> BinaryResult<u64> {
+        Ok(self.cursor.get_ref().len() as u64)
     }
 }
